* Bundles/FLAC/FLAC.h           make it compile with non-fragile ABI         * Bundles/Musepack/Musepack.m           fix potentially uninitialized return value         * CynthiuneFadingTextField.h           float -> CGFloat transition         * CynthiunePauseButton.m         * MBResultsPanel.m         * PlaylistViewController.m         * Player.m           NS(U)Integer transitions, and fixing some NSLog           format strings           All this together makes the Player fading when start playing           a song work again on amd64
git-svn-id: svn://svn.savannah.nongnu.org/gap/trunk@2218 d25b4930-3292-e311-b074-0010c6de5da8

+++ user-apps/Cynthiune/CynthiunePauseButton.m
@@ -99,7 +99,7 @@
   SET (secondaryImage, image);
 }
 
-- (void) setState: (int) state
+- (void) setState: (NSInteger) state
 {
   if ([self isEnabled])
     {
